🎨 Put manpage generation to dedicated module
